+++ HISTORY.md
@@ -5,6 +5,7 @@
 ### 2020-07-27
 * threads: fixed binding contexts to other receivers (variable scope for ASK / TELL)
 * gui: updated "About Snap!" dialog box
+* gui: library browser: import selected library on pressing enter
 
 ### 2020-07-26
 * objects: fixed sprite speech balloon display for sounds